pyproject.toml
[tool.poetry]
name = "repository-assistant-service"
version = "0.2.8"
description = ""
authors = ["Eko Indarto <eko.indarto@dans.knaw.nl>"]
readme = "README.md"
include = ["resources/*", "logs/*"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.26.0"
dynaconf = "^3.2.3"
fastapi = "^0.109.0"
emoji = "^2.9.0"
gunicorn = "^21.2.0"
python-keycloak = "^4.2.2"
rule-engine = "^4.5.0"
jmespath = "^1.0.1"
json-logic = {git = "https://github.com/nadirizr/json-logic-py.git"}
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"